8271047: Implement 'Glass' modifications needed for Metal rendering pipeline

+++ modules/javafx.graphics/src/main/java/com/sun/glass/ui/win/WinView.java
@@ -66,7 +66,7 @@ final class WinView extends View {
     }
 
     @Override
-    protected int _getNativeFrameBuffer(long ptr) {
+    protected long _getNativeFrameBuffer(long ptr) {
         return 0;
     }
